为对接前端持续优化

+++ src/main/java/fun/tanc/selfclocking/service/UserTaskServiceImpl.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import fun.tanc.selfclocking.dao.UserDao;
 import fun.tanc.selfclocking.dao.UserTaskDao;
+import fun.tanc.selfclocking.model.Schedule;
 import fun.tanc.selfclocking.model.UserModel;
 import fun.tanc.selfclocking.model.UserTask;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,9 +39,6 @@ public class UserTaskServiceImpl {
             return false;
         }
 
-
-
-
         //更新用户task字段
         int insert = userTaskDao.insert(new UserTask(userID, taskName, taskStr));
         if (insert < 0) {
@@ -54,9 +52,12 @@ public class UserTaskServiceImpl {
     //删除任务字段
     public Boolean deleteUserTask(String userName,String taskName){
         UserModel userModel = usImpl.findUser(userName);
-        QueryWrapper<UserTask> eq = new QueryWrapper<UserTask>().eq("user_id", userModel.getId()).eq("task_name", taskName);
-        int delete = userTaskDao.delete(eq);
-        if (delete < 0) {
+        System.out.println(userModel);
+        System.out.println(userModel.getId());
+        System.out.println(taskName);
+        int delete = userTaskDao.delete(new QueryWrapper<UserTask>().eq("user_id", userModel.getId()).eq("task_name", taskName));
+        System.out.println(delete);
+        if (delete <= 0) {
             System.out.println("删除失败");
             return false;
         }
@@ -66,14 +67,32 @@ public class UserTaskServiceImpl {
     //列出所有任务字段
     public List<UserTask> finderUserTask(String username){
         UserModel userModel = usImpl.findUser(username);
-        return userTaskDao.selectList(new QueryWrapper<UserTask>().eq("user_id", userModel.getId()));
+        List<UserTask> userTaskList = userTaskDao.selectList(new QueryWrapper<UserTask>().eq("user_id", userModel.getId()));
+        if(userTaskList.isEmpty()){
+            return null;
+        }
+        return userTaskList;
     }
 
     //查询任务字段(模糊)
     public List<UserTask> findUserTask(String userName,String taskName){
         UserModel userModel = usImpl.findUser(userName);
-        List<UserTask> userTasks = userTaskDao.selectList(new QueryWrapper<UserTask>().eq("user_id", userModel.getId()).like("task_name", taskName));
-        userTasks.forEach(System.out::println);
-        return userTasks;
+        return userTaskDao.selectList(new QueryWrapper<UserTask>().eq("user_id", userModel.getId()).like("task_name", taskName));
+    }
+
+    //修改任务字段
+    public Boolean updateUserTask(String userName,String taskName,String taskStr){
+        UserModel userModel = usImpl.findUser(userName);
+        UserTask userTask = userTaskDao.selectOne(new QueryWrapper<UserTask>().eq("user_id", userModel.getId()).eq("task_name", taskName));
+        if  (userTask == null) {
+            System.out.println("任务不存在");
+            return null;
+        }
+        int update = userTaskDao.update(new UserTask(userModel.getId(), taskName, taskStr), new UpdateWrapper<UserTask>().eq("user_id", userModel.getId()).eq("task_name", taskName));
+        if (update < 0) {
+            System.out.println("修改失败");
+            return false;
+        }
+        return true;
     }
 }
